refactor(#17): move main screen component on 1 level down
