+++ src/app/App/MainScreen/StatusBar/StatusBar.css
